requirements.txt
fastapi==0.89.1
SQLAlchemy==1.4.46
uvicorn==0.20.0
psycopg2==2.9.5
pydantic[email]
passlib[bcrypt]
python-multipart
python-jose[cryptography]